// src/types/formData.ts
export interface FormData {
  name: string;
  type: string;
  status: string;
  address: string;
  unit: string;
  city: string;
  state: string;
  cep: string;
  price: string;
  condominium: string;
  iptu: string;
  internalArea: string;
  externalArea: string;
  lotArea: string;
  bedrooms: string;
  suites: string;
  bathrooms: string;
  garages: string;
  garageType: string;
  garageLocation: string;
  garageNumber: string;
  garageLocationDetails: string;
  keyLocation: string;
  keyLocationOther: string;
  description: string;
  features: string[];
  propertyStage: string;
  deliveryDate: string;
  furnished: string;
  typeFloor: string;
  unitDescription: string;
  condominiumLeisure: string;
  selectedCondominium: string;
  constructionCompany: string;
  constructionCompanyYears: string;
  ownerName: string;
  ownerPhone: string;
  capturer: string;
  paymentMethods: string;
  directWithOwner: string;
  acceptsFinancing: string;
  downPayment: string;
  installments: string;
  reinforcement: string;
  settlementValue: string;
}